WIP: Fix Run reading bugs
I think the remaining bug is that OVCs aren't right for runs read in from file, but not sure

+++ src/ExternalMergeSort.cpp
@@ -27,7 +27,7 @@ ExternalMergeSortIterator::ExternalMergeSortIterator(const ExternalMergeSortPlan
     while (r != nullptr) {
         CacheSizedRun *run = new CacheSizedRun();
         // Fill in the run array
-        for (int i = 0; i < RUN_RECORDS; i++) {
+        for (uint64_t i = 0; i < RUN_RECORDS; i++) {
             run->push(r);
             r = _input->next();
         }
